+++ bindings/python/src/expose-problem.cpp
@@ -4,6 +4,7 @@
 #include "aligator/core/traj-opt-problem.hpp"
 #include "aligator/core/traj-opt-data.hpp"
 #include "aligator/core/cost-abstract.hpp"
+#include <eigenpy/deprecation-policy.hpp>
 
 namespace aligator {
 namespace python {
@@ -18,20 +19,23 @@ void exposeProblem() {
   using context::TrajOptProblem;
   using context::UnaryFunction;
 
-  using PolyFunction = xyz::polymorphic<UnaryFunction>;
+  using PolyUnaryFunction = xyz::polymorphic<UnaryFunction>;
+  using PolyFunction = xyz::polymorphic<context::StageFunction>;
   using PolyStage = xyz::polymorphic<StageModel>;
   using PolyCost = xyz::polymorphic<CostAbstract>;
   using PolyManifold = xyz::polymorphic<Manifold>;
+  using PolySet = xyz::polymorphic<context::ConstraintSet>;
 
   bp::class_<TrajOptProblem>("TrajOptProblem", "Define a shooting problem.",
                              bp::no_init)
-      .def(bp::init<PolyFunction, const std::vector<PolyStage> &, PolyCost>(
-          "Constructor adding the initial constraint explicitly.",
-          ("self"_a, "init_constraint", "stages", "term_cost")))
+      .def(
+          bp::init<PolyUnaryFunction, const std::vector<PolyStage> &, PolyCost>(
+              "Constructor adding the initial constraint explicitly.",
+              ("self"_a, "init_constraint", "stages", "term_cost")))
       .def(bp::init<ConstVectorRef, const std::vector<PolyStage> &, PolyCost>(
           "Constructor for an initial value problem.",
           ("self"_a, "x0", "stages", "term_cost")))
-      .def(bp::init<PolyFunction, PolyCost>(
+      .def(bp::init<PolyUnaryFunction, PolyCost>(
           "Constructor adding the initial constraint explicitly (without "
           "stages).",
           ("self"_a, "init_constraint", "term_cost")))
@@ -54,9 +58,21 @@ void exposeProblem() {
                     &TrajOptProblem::setInitState, "Initial state.")
       .add_property("init_constraint", &TrajOptProblem::init_constraint_,
                     "Get initial state constraint.")
-      .def("addTerminalConstraint", &TrajOptProblem::addTerminalConstraint,
-           ("self"_a, "constraint"), "Add a terminal constraint.")
+      .def<void (TrajOptProblem::*)(const context::StageConstraint &)>(
+          "addTerminalConstraint", &TrajOptProblem::addTerminalConstraint,
+          eigenpy::deprecated_member<>("This method is deprecated (because "
+                                       "StageConstraint has been deprecated)."),
+          ("self"_a, "constraint"), "Add a terminal constraint.")
+      .def<void (TrajOptProblem::*)(const PolyFunction &, const PolySet &)>(
+          "addTerminalConstraint", &TrajOptProblem::addTerminalConstraint,
+          ("self"_a, "func", "set"), "Add a terminal constraint.")
       .def("removeTerminalConstraint",
+           &TrajOptProblem::removeTerminalConstraints,
+           eigenpy::deprecated_member<>(
+               "This method is deprecated (due to a typo which was fixed). Use "
+               "removeTerminalConstraints instead."),
+           ("self"_a), "Remove all terminal constraints.")
+      .def("removeTerminalConstraints",
            &TrajOptProblem::removeTerminalConstraints, "self"_a,
            "Remove all terminal constraints.")
       .def("evaluate", &TrajOptProblem::evaluate,